% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calc_discharge_anomalies}
\alias{calc_discharge_anomalies}
\title{Calculate current, 7-, 14- and 28- day percentiles of discharge at USGS stations.
Uses \code{furrr} to process statistics in parallel. By default,
\code{furrr} uses a sequential plan to get discharge, but you can specify a multisession
plan by running \code{future::plan(future::multisession, workers = 5)} before running the code.}
\usage{
calc_discharge_anomalies(discharge)
}
\arguments{
\item{discharge}{The \code{sf} object returned by \code{\link[=get_discharge_shp]{get_discharge_shp()}}.}
}
\value{
The same \code{sf} as was input, but with a \code{time} column corresponding to
the time period of aggregation, a \code{value} column corresponding to the percentile
value of the observation, and a \code{fillColor} column corresponding to the USDM color
palette for the percentile.
}
\description{
Calculate current, 7-, 14- and 28- day percentiles of discharge at USGS stations.
Uses \code{furrr} to process statistics in parallel. By default,
\code{furrr} uses a sequential plan to get discharge, but you can specify a multisession
plan by running \code{future::plan(future::multisession, workers = 5)} before running the code.
}
\examples{
\dontrun{
stations = get_gauges()
discharge = get_discharge_shp(stations)
percentiles = calc_discharge_anomalies(discharge)
}
}
