% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domain}
\alias{domain}
\title{An \code{sf} object that is an example domain for clipping USGS data}
\format{
A \code{sf::sf} object of the Upper Missouri River Basin
\describe{
\item{CRS}{EPSG 4326}
}
}
\usage{
domain
}
\description{
An \code{sf} object that is an example domain for clipping USGS data
}
\keyword{datasets}
