% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_discharge.R
\name{get_discharge}
\alias{get_discharge}
\title{Get the last 30 years of discharge data from a given USGS Gauge.}
\usage{
get_discharge(STAID)
}
\arguments{
\item{STAID}{The 8-digit USGS station code to get data for.}
}
\value{
A \code{tibble} with "date" and "val" columns, where "val" is the discharge in cfs.
}
\description{
Get the last 30 years of discharge data from a given USGS Gauge.
}
\examples{
get_discharge("05014500")
}
