% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gauges.R
\name{get_gauges}
\alias{get_gauges}
\title{Get an \code{sf} object of active USGS stream gauges.}
\usage{
get_gauges(clip_shp = NULL)
}
\arguments{
\item{clip_shp}{An optional \code{sf} object to clip the output to.}
}
\value{
An \code{sf} object clipped to \code{clip_shp} of USGS gauges.
}
\description{
Get an \code{sf} object of active USGS stream gauges.
}
\examples{
\dontrun{
gauges <- get_stations_shp()
}
}
